requirements.txt
requests
requests-cache
Pillow
customtkinter
